Refactor API endpoints
