--- src/rust/examples/web-server/src/request/mod.rs
@@ -1,3 +1,0 @@
-mod objects;
-
-pub use objects::*;
