--- src/rust/examples/web-server/src/request/objects.rs
@@ -1,15 +1,0 @@
-#[derive(Deserialize, Serialize)]
-#[serde(rename_all="camelCase")]
-pub struct PayloadRegister {
-    pub public_key: String,
-    pub timestamp: String,
-    pub entered_text: String,
-}
-
-#[derive(Deserialize, Serialize)]
-#[serde(rename_all="camelCase")]
-pub struct PayloadCoolStuff {
-    pub timestamp: String,
-    pub coolness: String,
-    pub uuid: String,
-}
