+++ src/rust/examples/web-server/src/response/objects.rs
@@ -3,16 +3,3 @@ pub struct Error {
     pub status: u16,
     pub message: String,
 }
-
-#[derive(Serialize)]
-#[serde(rename_all="camelCase")]
-pub struct Register {
-    pub uuid: String,
-    pub welcome_message: String,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all="camelCase")]
-pub struct CoolStuff {
-    pub double_cool: String,
-}
